Add Bun ORM support for DB2 dialect (#284)
* Bun Support for DB2 Dialect


* Fixed pull review comments


* Some more fixes


* Fixed some connection pool issues


* Fixed some more code review issues


* Fixed some more code review issues


* Fixed some more code review issues


* Added testcase for sample bun run


* Bun DB2 dialect support for LUW and z/OS


* Fix DB2 parameter binding and platform-specific tests

Use date/time-specific ODBC bindings, avoid unsupported z/OS parameter description for SELECT statements, and gate DB2 platform-incompatible integration tests.


* Split Bun ORM integration into separate db2dialect module to keep core driver dependency-free


* code review changes


---------

+++ column.go
@@ -155,7 +155,7 @@ func (c *BaseColumn) TypeScan() reflect.Type {
 	case api.SQL_C_BIT:
 		return reflect.TypeOf(false)
 	case api.SQL_C_LONG:
-		return reflect.TypeOf(int32(0))
+		return reflect.TypeOf(int64(0))
 	case api.SQL_C_SBIGINT:
 		return reflect.TypeOf(int64(0))
 	case api.SQL_C_DOUBLE:
@@ -185,7 +185,8 @@ func (c *BaseColumn) Value(buf []byte) (driver.Value, error) {
 	case api.SQL_C_BIT:
 		return buf[0] != 0, nil
 	case api.SQL_C_LONG:
-		return *((*int32)(p)), nil
+		// driver.Value only allows int64 (not int32), per database/sql/driver contract
+		return int64(*((*int32)(p))), nil
 	case api.SQL_C_SBIGINT:
 		return *((*int64)(p)), nil
 	case api.SQL_C_DOUBLE:
